Restructure whole project - introduce packages
